+++ src/fieldtypes/ProductFieldType.php
@@ -14,8 +14,11 @@ class ProductFieldType extends Field implements PreviewableFieldInterface
      * @param $value
      * @return mixed
      */
-    public function normalizeValue($value, ElementInterface $element = null) {
-        if(is_array($value)) return $value;
+    public function normalizeValue($value, ElementInterface $element = null)
+    {
+        if (is_array($value)) {
+            return $value;
+        }
         return json_decode($value);
     }
 
@@ -32,7 +35,7 @@ class ProductFieldType extends Field implements PreviewableFieldInterface
         //category is the filename inside ./translations/en/
         return Craft::t('shopify', 'Shopify Product');
     }
-    
+
     /**
      * returns the template-partial an editor sees when editing plugin-content on a page
      *
@@ -44,25 +47,50 @@ class ProductFieldType extends Field implements PreviewableFieldInterface
      */
     public function getInputHtml($value, ElementInterface $element = null): string
     {
-        $products = Shopify::getInstance()->service->getProducts();
+        $defaultOptions = [];
+        $count = Shopify::getInstance()->service->getProductsCount($defaultOptions);
+        $productsData = Shopify::getInstance()->service->getProducts($defaultOptions);
+
+        $products = [];
+        if ($productsData['products'] && count($productsData['products']) > 0) {
+            $products = array_merge_recursive($products, $productsData['products']);
+        }
+
+        if (count($products) < $count && $productsData['link']['url']) {
+            while (count($products) < $count) {
+                $nextLink = $productsData['link']['url'];
+                $productsData = Shopify::getInstance()->service->getProducts($defaultOptions, $nextLink);
+                if ($productsData['products'] && count($productsData['products']) > 0) {
+                    $products = array_merge_recursive($products, $productsData['products']);
+                }
+            }
+        }
 
         $options = [];
-        if($products) {
+        if ($products) {
             foreach ($products as $product) {
                 $options[] = array(
                     'label' => $product['title'],
-                    'productId' => $product['id']
+                    'productId' => $product['id'],
+                    'sku' => implode(
+                        ', ',
+                        array_map(function ($variant) {
+                            return $variant['sku'];
+                        }, $product['variants'])
+                    ),
                 );
             }
         }
 
-        return Craft::$app->getView()->renderTemplate('shopify/_select',
-            [
-                'name' => $this->handle,
-                'value' => $value,
-                'field' => $this,
-                'products' => $options
-            ]);
+        return Craft::$app->getView()->renderTemplate('shopify/_select', [
+            'filter_class' => $this->handle . '_filter',
+            'selected_only_class' => $this->handle . '_selected_only',
+            'clear_selected_class' => $this->handle . '_clear_selected',
+            'wrapper_class' => $this->handle . '_wrapper',
+            'name' => $this->handle,
+            'value' => $value,
+            'field' => $this,
+            'products' => $options,
+        ]);
     }
-
 }