+++ CHANGELOG.md
@@ -1,21 +1,40 @@
-# shopify Changelog
+# Shopify Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.0.5 - 2019-10-10
+## 1.1.0 - ???
+
 ### Fixed
-- GitHub-issue #7
+
+-   GitHub-issue #8
+
 ### Added
-- data-normalization to use selected values as an array
+
+-   Allow limit and published_status to be configured in the settings menu.
+-   UI tweaks for the default field that allow more filtering and display tweaks.
+
+## 1.0.5 - 2019-10-10
+
+### Fixed
+
+-   GitHub-issue #7
+
+### Added
+
+-   data-normalization to use selected values as an array
 
 ## 1.0.4 - 2019-10-07
+
 ### Fixed
-- Fixed an issue with class-path for Craft3-Plugin class-name
-- removed -RC1 flag of craftcms dependency
-- use correct craft input-multiselect classes
+
+-   Fixed an issue with class-path for Craft3-Plugin class-name
+-   removed -RC1 flag of craftcms dependency
+-   use correct craft input-multiselect classes
 
 ## 1.0.0 - 2018-07-22
+
 ### Added
-- Initial release
+
+-   Initial release